+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add VCC voltage and hardware revision to info screen ([#143](https://github.com/OpenRemise/Frontend/pull/143))
 - Blue power off icon indicates short circuit ([#143](https://github.com/OpenRemise/Frontend/pull/143))
 - Allow smaller text scaling (0.8) ([#158](https://github.com/OpenRemise/Frontend/issues/158))
+- Add OpenDCC `.Locxml` import/export ([#172](https://github.com/OpenRemise/Frontend/issues/172))
 - Bugfix higher functions F31-F13 from `LAN_X_LOCO_INFO` incorrectly shifted ([#162](https://github.com/OpenRemise/Frontend/issues/162))
 - Bugfix `z21ServiceProvider` must use correct API for provider self invalidation ([#163](https://github.com/OpenRemise/Frontend/issues/163))
 - Bugfix exception in settings screen ([#166](https://github.com/OpenRemise/Frontend/issues/166))